Added a new result (error) as not leader, deleted the duplicate port defining, make sure to avoid the warnings with clarifying the type of the cluster size (int)

+++ core/raft.c
@@ -369,7 +369,7 @@ void raft_tick(void)
                 }
             }
 
-            if (votes > (g_cluster_size / 2U)) {
+            if (votes > (int)(g_cluster_size / 2U)) {
                 g_raft.commit_idx = n;
                 break;
             }
